refactor(package-editor): consolidate package editor components and hooks
Moved package editor components and hooks to a centralized location for better maintainability. Deleted redundant files and updated imports accordingly. This refactor improves code organization and reduces duplication.
