+++ src/App.tsx
@@ -73,19 +73,13 @@ const DevLoginPage = lazyImport(() => import("@/pages/dev-login"))
 const BetaPage = lazyImport(() => import("@/pages/beta"))
 const ViewPackagePage = lazyImport(() => import("@/pages/view-package"))
 const TrendingPage = lazyImport(() => import("@/pages/trending"))
-
-// Wip Package Port Routes
 const PackageEditorPage = lazyImport(async () => {
   const [editorModule] = await Promise.all([
-    import("@/pages/p/editor"),
+    import("@/pages/package-editor"),
     import("@/lib/utils/load-prettier").then((m) => m.loadPrettier()),
   ])
   return editorModule
 })
-const PackageUserProfilePage = lazyImport(
-  () => import("@/pages/p/user-profile"),
-)
-const PackageViewPage = lazyImport(() => import("@/pages/p/view-package"))
 
 class ErrorBoundary extends React.Component<
   { children: React.ReactNode },
@@ -114,11 +108,6 @@ function App() {
       <ErrorBoundary>
         <Suspense fallback={<FullPageLoader />}>
           <Switch>
-            {/* Package Wip Routes  */}
-            <Route path="/p/editor" component={PackageEditorPage} />
-            <Route path="/p/:username" component={PackageUserProfilePage} />
-            <Route path="/p/:author/:packageName" component={PackageViewPage} />
-
             <Route path="/" component={LandingPage} />
             <Route path="/beta" component={BetaPage} />
             <Route path="/beta/:author/:packageName" component={BetaPage} />
@@ -126,6 +115,7 @@ function App() {
               path="/view-package/:author/:packageName"
               component={ViewPackagePage}
             />
+            <Route path="/p/editor" component={PackageEditorPage} />
             <Route path="/editor" component={EditorPage} />
             <Route path="/quickstart" component={QuickstartPage} />
             <Route path="/dashboard" component={DashboardPage} />
